Add logic to users, incomes, paymentMethot and fix bags

+++ Entities/DTOs/Response/ExpenseResponseDto.cs
@@ -1,11 +1,10 @@
 ﻿using ExpenseTracker.Entities.Models;
-using ExpenseTracker.Enums;
 
 namespace ExpenseTracker.Entities.DTOs.Response
 {
     public class ExpenseResponseDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; } = string.Empty;
         public decimal Amount { get; set; }
         public int CategoryId { get; set; }
